Rename class.
