+++ xbird-starter-id/src/main/java/cn/xbird/starter/id/annotations/EnableSegment.java
@@ -16,7 +16,6 @@
  */
 package cn.xbird.starter.id.annotations;
 
-import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -24,16 +23,13 @@ import java.lang.annotation.Target;
 
 import org.springframework.context.annotation.Import;
 
-import cn.xbird.starter.id.boot.SegmentConfiguration;
-
 /**
  * @author zhycn
  * @since 2.2.0 2018-06-15
  */
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
-@Documented
-@Import(SegmentConfiguration.class)
+@Import(SegmentConfigurationSelector.class)
 public @interface EnableSegment {
 
 }
